==> 01_getStatistics.ipynb <==
import pandas as pd

df = pd.read_csv('herg_compound_from_chembl.csv', index_col=[0])
df.head()

df.columns

df.activity_comment.value_counts()

df.assay_type.value_counts()

df.standard_relation.value_counts()

df.units.value_counts()

import numpy as np

def unitConversion(value, unit='uM'):
    if unit == 'uM':
        value = float(value) * 1000
    elif unit == 'nM' or unit == '\nM':
        value = float(value)
    elif unit == 'microM':
        value = float(value) * 1000000
    else:
        pass
    return value

df.value = [unitConversion(x,df.units[i]) for i,x in enumerate(df.value)]
df.units = 'nM'

df.value = [float(x) for x in df.value]
df = df[df.activity_comment != 'Not Determined']
df = df[df.activity_comment != 'Note: corresponding IC50 reported as Active']
df = df[df.assay_type == 'B'] # 只保留Binding Data

print('共有%s个化合物' %(len(df)))
print('活性化合物(Ki<10uM)有%s个' %len(df[df.value < 10000]))
print('非活性化合物(Ki<10uM)有%s个' %len(df[(df.value >= 10000) | (df.activity_comment == 'Not Active (inhibition < 50% @ 10 uM and thus dose-reponse curve not measured)') | (df.activity_comment == 'Not Active')]))

# activity_comment列标记活性
df.activity_comment[df.value < 10000] = 1
df.activity_comment[(df.value >= 10000) | (df.activity_comment == 'Not Active (inhibition < 50% @ 10 uM and thus dose-reponse curve not measured)') | (df.activity_comment == 'Not Active')] = 0

df.to_csv('herg_compound_from_chembl_processed.csv')
