==> 02_keepLargestFrag.ipynb <==
from rdkit import Chem
import pandas as pd

df = pd.read_csv('herg_compound_from_chembl_processed.csv', index_col=[0])
df.columns

print(len(df))
df.canonical_smiles.fillna(0, inplace=True)
df = df[df.canonical_smiles != 0]
df.mol = [Chem.MolFromSmiles(x) for x in df.canonical_smiles]
print(len(df))

largest_frag_lists = []
for s in df.mol:
    fragmented = Chem.GetMolFrags(s, asMols=True)
    list_cpds_fragsize = []
    for x in fragmented:
        list_cpds_fragsize.append(x.GetNumAtoms())
    largest_frag_index = list_cpds_fragsize.index(max(list_cpds_fragsize))
    largest_frag = fragmented[largest_frag_index]
    largest_frag_lists.append(Chem.MolToSmiles(largest_frag))
    
df['smiles'] = largest_frag_lists

df.to_csv('herg_compound_from_chembl_processed_largestFrag.csv', index=False)
