==> 03_delDupl.ipynb <==
from rdkit import Chem
from rdkit.Chem.Fingerprints import FingerprintMols
import pandas as pd

df = pd.read_csv('herg_compound_from_chembl_processed_largestFrag.csv')

df['Mol'] = [Chem.MolFromSmiles(x) for x in df.smiles]
df['Mol'] = [Chem.RemoveHs(x) for x in df.Mol]
df['fp'] = [FingerprintMols.FingerprintMol(x) for x in df.Mol]
df = df.drop_duplicates(subset=['fp'],keep='first')

print(len(df))

df.to_csv('herg_compound_from_chembl_processed_largestFrag_delDupl.csv', index=False)
