==> 04_calc_descr.ipynb <==
from rdkit import Chem
from rdkit.Chem import Descriptors
from rdkit.ML.Descriptors import MoleculeDescriptors
import pandas as pd
import tqdm

df = pd.read_csv('herg_compound_from_chembl_processed_largestFrag_delDupl.csv')

df.Mol = [Chem.MolFromSmiles(x) for x in df.smiles]

nms = [x[0] for x in Descriptors._descList if x[0] != 'MolecularFormula']
print(len(nms))
calc = MoleculeDescriptors.MolecularDescriptorCalculator(nms)

for nam in nms:
    df[nam] = 0
    
for i, s in enumerate(df.Mol[:]):
    descrs = calc.CalcDescriptors(s)
#     print(list(descrs))
#     for j, x in enumerate(descrs):
    df.loc[i, nms] = descrs

# list(df.columns)

s_list = ['smiles'] + nms + ['activity_comment']
df[s_list].to_csv('herg_descr.csv', index=False)
